+++ lib/yandex_login_sdk_platform_interface.dart
@@ -1,6 +1,7 @@
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 import 'src/yandex_login_result.dart';
+import 'src/yandex_login_strategy.dart';
 import 'yandex_login_sdk_method_channel.dart';
 
 abstract class YandexLoginSdkPlatform extends PlatformInterface {
@@ -17,7 +18,10 @@ abstract class YandexLoginSdkPlatform extends PlatformInterface {
     _instance = instance;
   }
 
-  Future<YandexLoginResult> signIn({required String clientId}) {
+  Future<YandexLoginResult> signIn({
+    required String clientId,
+    YandexLoginStrategy strategy = YandexLoginStrategy.auto,
+  }) {
     throw UnimplementedError('signIn() has not been implemented.');
   }